fix Hashtbl build
The build failure was as follows:

    Error: The implementation extHashtbl.ml
       does not match the interface extHashtbl.cmi:
       ...
       At position module Hashtbl : sig module type S = <here> end
       The value `find_opt' is required but not provided

The error comes from the fact that extHashtbl.ml uses `include
Hashtbl`, including the `Hashtbl` module of the standard library,
including its `S` and `SeededS` signature, while extHashtbl.mli
redefines/duplicates those signatures. As `find_opt` was added to the
standard library, the included and copied signatures did not match
anymore -- note that there is no subtyping possible in this case, if
a module and its implementation define a module signature, they have
to define exactly the same signature.